Fix windows build
